requirements.txt
flask
flask-sqlalchemy
werkzeug
python-dotenv
google-generativeai
